refactor(earn): extract ApyWindow types to neutral module, fix imports
